+++ NEWS.md
@@ -1,3 +1,5 @@
+# guideR (development version)
+
 # guideR 0.4.1
 
 **New features**