+++ NEWS.md
@@ -1,12 +1,15 @@
 # guideR (development version)
 
-# guideR 0.4.1
-
 **New features**
 
 * new function `plot_multiple_answers()` to plot a multiple answers question
   coded as several binary variables (#29)
 * new helper `combine_answers()` for multiple answers questions (#29)
+
+# guideR 0.4.1
+
+**New feature**
+
 * new argument `dependencies` for `install_dependencies()` (#27)
 
 # guideR 0.4.0